pyproject.toml
[project]
name = "pg-analytics-mcp"
version = "0.2.0"
description = "MCP server for PostgreSQL analytics — schema discovery, data exploration, relationships, performance, data quality, multi-env support"
license = {text = "MIT"}
keywords = ["mcp", "postgresql", "analytics", "data-quality", "monitoring"]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.0",
    "psycopg2-binary>=2.9",
]

[project.urls]
Homepage = "https://github.com/fabdendev/pg-analytics-mcp"
Repository = "https://github.com/fabdendev/pg-analytics-mcp"
Issues = "https://github.com/fabdendev/pg-analytics-mcp/issues"

[project.scripts]
pg-analytics-mcp = "pg_analytics_mcp.server:main"

[project.optional-dependencies]
dev = ["pytest>=9.0", "ruff>=0.14"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["pg_analytics_mcp"]

[tool.pytest.ini_options]
pythonpath = ["."]
